==> 1.FIFA_WC_ML_Investigation.ipynb <==
import pandas as pd
from string import ascii_uppercase as alphabet
import pickle

todas_tablas = pd.read_html('https://en.wikipedia.org/wiki/2022_FIFA_World_Cup')

# **Extracción de todos los Grupos y convirtiendolos en un archivo pickle**

todas_tablas[9]
todas_tablas[16]
todas_tablas[23]
todas_tablas[30]
todas_tablas[37]
todas_tablas[44]
todas_tablas[51]
todas_tablas[58]

alphabet

for word, i in zip(alphabet, range(9, 65, 7)):
  print(word, i)

dict_tablas = {}
for word, i in zip(alphabet, range(9, 65, 7)):
  df = todas_tablas[i]
  df.rename(columns={df.columns[1]:'Team'}, inplace=True)

  if 'Qualification' in df.columns:
    df.pop('Qualification')
  dict_tablas[f'Grupo {word}'] = df

dict_tablas.keys()

dict_tablas['Grupo B']

# para exportar el diccionario
from google.colab import drive
drive.mount('/content/drive', force_remount=True)

path = '/content/drive/MyDrive/DataScience/2. Proyecto de Data Science con Python - YT/dict_table.pickle'

with open(path, 'wb') as file:
  pickle.dump(dict_tablas, file)

# **Extracción de los partidos jugados en la fase de grupos, octavos de final, semifinal y la final utilizando Beatiful Soup**

!pip install bs4
