==> 4.FIFA_WC_ML_DataCleaning.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np

## **1. Cargando los Datasets con los que vamos a trabajar**

dir_datasets = '/content/drive/MyDrive/DataScience/2. Proyecto de Data Science con Python - YT/0.Dataset/{}'

df_data_historica = pd.read_csv(dir_datasets.format('fifa_worldcup_matches.csv'))
df_fixture = pd.read_csv(dir_datasets.format('fifa_worldcup_fixture.csv'))
df_data_faltante = pd.read_csv(dir_datasets.format('fifa_worldcup_missing_data.csv'))

# **2. Limpiando df_fixture**

# vamos a limpiar las columnas home y away de los espacios en blanco
df_fixture['home'] = df_fixture['home'].str.strip()
df_fixture['away'] = df_fixture['away'].str.strip()

# **3. Limpiando df_missing_data y concatenar al dataframe orginal(df_data_historica)**

**3.1 Revisamos si hay data faltante en las columnas y luego eliminamos**

df_data_faltante[df_data_faltante['home'].isnull()]
df_data_faltante.dropna(inplace=True)

**3.2 Ahora vamos a concatenar el dataframe orginal historica con el dataframe faltante**

df_data_historica = pd.concat([df_data_historica, df_data_faltante], ignore_index=True)
df_data_historica

**3.3 Ahora vamos a eliminar duplicados que podria tener un dataframe y a ordenarlo**

df_data_historica.drop_duplicates(inplace=True)
df_data_historica.sort_values('year', inplace=True)
df_data_historica

**3.4 Hay un partido que no se desputo, por lo cual no tiene marcador, entonces hay que eliminarlo**

df_data_historica[df_data_historica['score']=='w/o[a]'] 

**3.5 Vamos a eliminar una fila que no tiene marcador, porque uno de los dos equipos no se presento**

index_eliminar = df_data_historica[df_data_historica['score']=='w/o[a]'].index

**3.6 Eliminamos la fila del partido que no se jugo**

df_data_historica.drop(index = index_eliminar, inplace=True)

**3.7 Ahora seguimos buscando en la columna score que no contenga digitos**

df_data_historica[df_data_historica['score'].str.contains('[^\d–]')]

**3.8 Ahora vamos a eliminar (a.e.t.) que esta dentro de algunas filas, con el metodo replace**

df_data_historica['score'] = df_data_historica['score'].str.replace('[^\d–]', '', regex=True)

**3.9 Verificamos que se reeplacen**

df_data_historica[df_data_historica['score'].str.contains('[^\d–]')]

**3.10 Ahora limpiamos las columnas home y away data historica**

df_data_historica['home'] = df_data_historica['home'].str.strip()
df_data_historica['away'] = df_data_historica['away'].str.strip()

**3.11 Ahora vamos a separar el score en dos columnas, vamos a tener los goles en diferentes columnas**

df_data_historica[['HomeGoals', 'AwayGoals']] = df_data_historica['score'].str.split('–', expand=True)

**3.12 Ahora vamos a eliminar la columna score**

df_data_historica.drop('score', axis=1, inplace=True)

df_data_historica

**3.13 Ahora vamos a renombrar las columnas del DataFrame**

df_data_historica.rename(columns={'home': 'HomeTeam',
                                  'away': 'AwayTeam',
                                  'year': 'Year'}, inplace=True)

df_data_historica

**3.14 Ahora debemos cambiar el tipo de dato de las columnas HomeGoals y AwayGoals**

df_data_historica.dtypes

df_data_historica = df_data_historica.astype({'HomeGoals': int, 'AwayGoals': int, 'Year': int})

df_data_historica.dtypes

**3.15 Ahora vamos agregar una nueva columna para tener la cantidad de goles totales por partido**

df_data_historica['TotalGoals'] = df_data_historica['HomeGoals'] + df_data_historica['AwayGoals']
df_data_historica

# **4. Exportamos los DataFrames a un CSV con la informacion ya limpia**

# Exportar el DataFrame a un archivo CSV
df_data_historica.to_csv('/content/drive/MyDrive/DataScience/2. Proyecto de Data Science con Python - YT/0.Dataset/cleaned_fifa_worldcup_matches.csv', index=False)
df_fixture.to_csv('/content/drive/MyDrive/DataScience/2. Proyecto de Data Science con Python - YT/0.Dataset/cleaned_fifa_worldcup_fixture.csv', index=False)


**4.1 Verificar los números de partidos**

years = [1930, 1934, 1938, 1950, 1954, 1958, 1962, 1966, 1970, 1974,
         1978, 1982, 1986, 1990, 1994, 1998, 2002, 2006, 2010, 2014,
         2018]

for year in years:
    print(year, ' jugaron:', len(df_data_historica[df_data_historica['Year']==year]))
