==> 5.FIFA_WC_ML_Predict_Model.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import pickle 
from scipy.stats import poisson

dir_datasets = '/content/drive/MyDrive/DataScience/2. Proyecto de Data Science con Python - YT/0.Dataset/{}'
dir_dict = '/content/drive/MyDrive/DataScience/2. Proyecto de Data Science con Python - YT/dict_table.pickle'

dict_table = pickle.load(open(dir_dict, 'rb'))
df_data_historical = pd.read_csv(dir_datasets.format('cleaned_fifa_worldcup_matches.csv'))
df_fixture = pd.read_csv(dir_datasets.format('cleaned_fifa_worldcup_fixture.csv'))

# revisando los archivos
dict_table.keys()
dict_table['Group A']
df_data_historical.head()

# **1.Calcular Team Strenth**

# dividir df en df_home y df_away
df_home = df_data_historical[['HomeTeam', 'HomeGoals', 'AwayGoals']]
df_away = df_data_historical[['AwayTeam', 'HomeGoals', 'AwayGoals']]

# renombrar columnas
df_home = df_home.rename(columns={'HomeTeam': 'Team',
                        'HomeGoals': 'GoalsScored',
                        'AwayGoals': 'GoalsConceded'})

df_away = df_away.rename(columns={'AwayTeam': 'Team',
                        'HomeGoals': 'GoalsConceded',
                        'AwayGoals': 'GoalsScored'})

# concatenar df_home y df_away, hacer group por team y calcular promedio
df_team_strength = pd.concat([df_home, df_away], ignore_index=True).groupby('Team').mean()
df_team_strength

from google.colab import files

df_team_strength.to_csv('df_team_strength.csv', index=False)
files.download('df_team_strength.csv')

# **2.Función predict_points()**

def predict_points(home, away):
  if home in df_team_strength.index and away in df_team_strength.index:
    lamb_home = df_team_strength.at[home, 'GoalsScored'] * df_team_strength.at[away,'GoalsConceded']
    lamb_away = df_team_strength.at[away,'GoalsScored'] * df_team_strength.at[home,'GoalsConceded']
    prob_home, prob_away, prob_draw = 0, 0, 0
    #print(lamb_home)
    #print(lamb_away)
    for x in range(0, 11):
      for y in range(0, 11):
        p = poisson.pmf(x, lamb_home) * poisson.pmf(y, lamb_away)
        if x == y:
          prob_draw += p
        elif x > y:
          prob_home += p
        else:
          prob_away += p

    points_home = 3 * prob_home + prob_draw
    points_away = 3 * prob_away + prob_draw

    
    return (points_home, points_away)
  else:
    return (0, 0)

#predict_points('Argentina', 'Mexico')

**2.1 Testear mi funcion con:
Argentina - Mexico, Inglaterra - USA, Qatar - Ecuador**

print(predict_points('Argentina', 'Mexico'))
print(predict_points('England', 'United States'))
print(predict_points('Qatar (H)', 'Ecuador'))

# **3. Prediciendo el mundial**

**3.1 Prediciendo la fase de grupos**

# diviendo el fixture en grupos, octavos, cuartos, semis, etc
df_fixture_group_48 = df_fixture[:48].copy()
df_fixture_knockout = df_fixture[48:56].copy()
df_fixture_quarter = df_fixture[56:60].copy()
df_fixture_semi = df_fixture[60:62].copy()
df_fixture_final = df_fixture[62:].copy()

# correr todos los partidos de la fase de grupos y actualizar las tablas de cada grupo
for group in dict_table:
  teams_in_group = dict_table[group]['Team'].values
  df_fixture_group_6 = df_fixture_group_48[df_fixture_group_48['home'].isin(teams_in_group)]
  for index, row in df_fixture_group_6.iterrows():
      home, away = row['home'], row['away']
      points_home, points_away = predict_points(home, away)
      dict_table[group].loc[dict_table[group]['Team'] == home, 'Pts'] += points_home
      dict_table[group].loc[dict_table[group]['Team'] == away, 'Pts'] += points_away

  dict_table[group] = dict_table[group].sort_values('Pts', ascending=False).reset_index()
  dict_table[group] = dict_table[group][['Team', 'Pts']]
  dict_table[group] = dict_table[group].round(0)

# mostrar las tablas actualizadas
dict_table['Group A']
dict_table['Group B']
dict_table['Group C']
dict_table['Group D']
dict_table['Group E']
dict_table['Group F']
dict_table['Group G']
dict_table['Group H']

**3.2 Prediciendo los octavos de final**

df_fixture_knockout

dict_table['Group A']

# actualizar el fixture de octavos con el puesto 1 y 2 
for group in dict_table:
  group_winner = dict_table[group].loc[0, 'Team'] # primer puesto
  runner_up = dict_table[group].loc[1, 'Team'] # segundo puesto
  df_fixture_knockout.replace({f'Winners {group}' : group_winner,
                                       f'Runners-up {group}' : runner_up}, 
                              inplace=True)
  df_fixture_knockout['Winner'] = '?'

df_fixture_knockout

# obtener el ganador de los partidos de octavos
def get_winner(df_fixture_updated):
  for index, row in df_fixture_updated.iterrows():
    home, away = row['home'], row['away']
    points_home, points_away = predict_points(home, away)
    if points_home > points_away:
      winner = home
    else:
      winner = away

    df_fixture_updated.loc[index, 'Winner'] = winner
  
  return df_fixture_updated

# ejecutamos la funcion para que nos indique el ganador
get_winner(df_fixture_knockout)

**3.3 Prediciendo los cuartos de final**

# vamos a crear una funcion para cambiar los ganadores en los dataset correctos
def update_table(df_fixture_round_1, df_fixture_round_2): 
  for index, row in df_fixture_round_1.iterrows():
    winner = df_fixture_round_1.loc[index, 'Winner']
    matchh = df_fixture_round_1.loc[index, 'score']
    df_fixture_round_2.replace({f'Winners {matchh}':winner}, inplace=True)
  df_fixture_round_2['Winner'] = '?'
  return df_fixture_round_2

update_table(df_fixture_knockout, df_fixture_quarter)

get_winner(df_fixture_quarter)

**3.4 Prediciendo semi finales**

update_table(df_fixture_quarter, df_fixture_semi)

get_winner(df_fixture_semi)

**3.5 Prediciendo la final**

update_table(df_fixture_semi, df_fixture_final)

get_winner(df_fixture_final)
